Remove DeploymentTester and rename InternalDeploymentTester
