+++ controller-server/src/test/java/com/yahoo/vespa/hosted/controller/maintenance/DeploymentIssueReporterTest.java
@@ -13,7 +13,7 @@ import com.yahoo.vespa.hosted.controller.application.ApplicationPackage;
 import com.yahoo.vespa.hosted.controller.application.Change;
 import com.yahoo.vespa.hosted.controller.application.TenantAndApplicationId;
 import com.yahoo.vespa.hosted.controller.deployment.ApplicationPackageBuilder;
-import com.yahoo.vespa.hosted.controller.deployment.InternalDeploymentTester;
+import com.yahoo.vespa.hosted.controller.deployment.DeploymentTester;
 import com.yahoo.vespa.hosted.controller.persistence.MockCuratorDb;
 import com.yahoo.vespa.hosted.controller.versions.VespaVersion;
 import org.junit.Before;
@@ -49,13 +49,13 @@ public class DeploymentIssueReporterTest {
             .upgradePolicy("canary")
             .build();
 
-    private InternalDeploymentTester tester;
+    private DeploymentTester tester;
     private DeploymentIssueReporter reporter;
     private MockDeploymentIssues issues;
 
     @Before
     public void setup() {
-        tester = new InternalDeploymentTester();
+        tester = new DeploymentTester();
         issues = new MockDeploymentIssues();
         reporter = new DeploymentIssueReporter(tester.controller(), issues, Duration.ofDays(1), new JobControl(new MockCuratorDb()));
     }
